agent/nicegui_agent/template/pyproject.toml
[project]
name = "template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nicegui>=2.19.0",
    "pytest-asyncio>=1.0.0",
    "pytest-selenium>=4.1.0",
]
